Edit display property and wrapping.
The default display property caused the square divs to
be added one per line, which was not the intended result.
To fix this issue, the display value was changed to flex which
caused the divs to be added horizontally across the page. However,
this only had the effect of creating one long horizontal line of divs.
Therefore, the flex-wrap property was adjusted to force items onto a new line.